requirements.txt
google-genai
python-dotenv
scikit-learn
sentence-transformers
numpy
faiss-cpu